Implements browser extension legacy injectior and implements wallet features.
The browser extension will now (on click) inject `window.ethereum` and then forward requests made to that over the HotOstrich channel.  This allows something like the iframe provider or the extension's provider to work with a page that only implements the legacy `window.ethereum` communication channel.

The extension now also functions as a full wallet, though it currently has no UI for actually presenting wallet selection to the user, so it is just hard coded with a wallet (ledger at the moment, though I think that may not actually work from an extension).